Colour maps: CORINE Land Cover, and QGIS exports
Land cover is the case where assigning colours by frequency is simply
wrong: every CORINE class has a colour the field reads fluently, and
inventing another one makes a familiar map unreadable. Styling by a
column of CLC codes now recognizes them and applies the official
palette, its class names, and the nomenclature order, with a checkbox to
fall back to the generic palette.

The table is generated from the legend shipped with the dataset, and the
same parser reads any QGIS colour-map export, so a dataset that ships
one can be styled by it without anything being added here. A map is only
offered when it covers at least 90% of the column's values: CORINE
colours on a column that merely looks like class codes would be a
confident lie.

This needed the bin ceiling raised from 16 to 64 — CLC has 47 classes,
and 31 of them would have landed in "other". Bin colours are resolved
CPU-side per chunk, so the cost is the hidden-bin mask widening from u16
to u64.

+++ src/map/renderer.rs
@@ -25,12 +25,12 @@ pub struct DrawStyle {
     /// Bitmask of style bins hidden from the map (legend toggles).
     /// Only honored while `bin_colors` is set: without data styling
     /// every chunk reports bin 0 and the mask would hide the layer.
-    pub hidden_bins: u16,
+    pub hidden_bins: u64,
 }
 
 impl DrawStyle {
     fn bin_hidden(&self, bin: u8) -> bool {
-        self.bin_colors.is_some() && self.hidden_bins & (1u16 << bin) != 0
+        self.bin_colors.is_some() && self.hidden_bins & (1u64 << bin) != 0
     }
 }
